Added charts to admin dashboard Fixed issues with notifications by creating a parallel component which mirrors the Flask models.py Added email alerts to clients and students Modified admin users view to filter for different user types Added facility to email all students (this is to be extended to other common groups) Added a login count and tracking of last login date Introduced common.py for shared functions (this is to be exapnded at the point that the app is refactored) Added further template functions into macros.html for common links (eg delete item) Completed the staff blueprint including a staff menu in the base.html template Added requirement for students to explicitly submit a bid. This allows checks for team composition, PM, etc. Added preview feature to manage bid page Enabled students to create projects Added functions to allow claients and staff to manage bids for their projects Added several pieces of information to staff and student home screens to show project/team status

+++ app/student/views.py
@@ -1,40 +1,64 @@
-from flask import render_template, flash, redirect, url_for, request
+from flask import render_template, flash, redirect, url_for, request, abort
 from flask_login import current_user, login_required
 from collections import Counter
 
 from sqlalchemy import and_, or_, text
 from sqlalchemy.orm.exc import NoResultFound
+from sqlalchemy.sql.expression import case
 
 from . import student
 from .forms import *
 from app.models import *
+from app.common import *
+import logging
+logger = logging.getLogger()
 
 
-@student.route('/student_dashboard')
+
+@student.route('/student/projects/list')
 @login_required
-def dashboard():
+def projects():
     skill = request.args.get("skill")
     company = request.args.get("company")
     project_domain = Domain.query.filter(Domain.name == 'project').first()
     statuses = Status.query.filter(and_(Status.domain_id == project_domain.id,
                                         or_(Status.name == 'Live',
                                             Status.name == 'Withdrawn',
+                                            Status.name == 'Taken',
                                             Status.name == 'Complete'))).with_entities(Status.id)
-    projects = Project.query.filter(Project.status_id.in_(statuses))
+
+    accepted = get_status('team', 'Accepted')
+    submitted = get_status('team', 'Submitted')
+    team_ids = [m.team_id for m in TeamMember.query.filter(TeamMember.user_id == current_user.id).all()]
+    accepted_for = [t.project_id for t in Team.query.filter(and_(Team.id.in_(team_ids), Team.status_id == accepted.id))]
+    submitted_for = [t.project_id for t in Team.query.filter(and_(Team.id.in_(team_ids), Team.status_id == submitted.id))]
+
+    projects = Project.query.filter(or_(Project.status_id.in_(statuses), Project.client_id == current_user.id)).\
+        order_by((Project.client_id == current_user.id).desc()).\
+        order_by(Project.title)
+
+    projects = sorted(projects, key=lambda x: (x.id not in accepted_for, x.id not in submitted_for))
+
     skills = []
+    bids = {}
     for project in projects:
-        skills += [(s.skill_required.id, s.skill_required.name) for s in project.skills_required]
-    skills = Counter(skills)
+        bids[project.id] = len(get_bids(project))
+        skills += [(s.skill_required.name, s.skill_required.id) for s in project.skills_required]
+    skills = sorted(Counter(skills).items(), key=lambda x: x[0][0].lower())
+
     companies = []
-    companies += [(p.client.company.id, p.client.company.name) for p in projects]
-    companies = Counter(companies)
-    return render_template('student/dashboard.html',
+    companies += [(p.client.company.name, p.client.company.id) for p in projects]
+    companies = sorted(Counter(companies).items(), key=lambda x: x[0][0].lower())
+    return render_template('student/projects/list.html',
                            projects=projects,
+                           bids=bids,
+                           accepted_for=accepted_for,
+                           submitted_for=submitted_for,
                            skills=skills,
                            skill=skill,
                            companies=companies,
                            company=company,
-                           title="Dashboard")
+                           title="Projects")
 
 
 @student.route('/profile/edit', methods=['GET', 'POST'])
@@ -55,7 +79,7 @@ def profile():
             db.session.add(skill_offered)
         db.session.commit()
         flash('Profile updated')
-        return redirect(url_for('student.dashboard'))
+        return redirect(url_for('student.projects'))
 
     if form.errors:
         for field in form.errors:
@@ -63,18 +87,6 @@ def profile():
                 flash(error, 'error')
 
     return render_template('student/profile/profile.html', user=user, form=form, title="Edit profile")
-
-
-@student.route('/projects')
-@login_required
-def projects():
-    project_domain = Domain.query.filter(Domain.name == 'project').first()
-    statuses = Status.query.filter(and_(Status.id == project_domain.id,
-                                        or_(Status.name == 'Live',
-                                            Status.name == 'Withdrawn',
-                                            Status.name == 'Complete')))
-    projects = Project.query.filter(Project.status in statuses).all()
-    return render_template('student/projects/projects.html', projects=projects, title="Projects")
 
 
 @student.route('/project/<int:id>')
@@ -86,12 +98,110 @@ def project(id):
     teams = Team.query.filter(Team.project_id == id)
     team_member = TeamMember.query.filter(and_(TeamMember.user_id == current_user.id,
                                                TeamMember.team_id.in_([t.id for t in teams]))).first()
+    transitions = Transition.query.filter(Transition.from_status_id == project.status_id, Transition.admin_only == False).all()
     return render_template('student/projects/project.html',
                            project=project,
                            interest=interest,
                            flag=flag,
                            team_member=team_member,
+                           transitions=transitions,
                            title='View project')
+
+
+@student.route('/add_project', methods=['GET', 'POST'])
+@login_required
+def add_project():
+    add_project = True
+
+    form = ProjectForm()
+    form.skills_required.choices = [(s.id, s.name) for s in Skill.query.order_by(text('name'))]
+    project_domain = Domain.query.filter(Domain.name == 'project').first()
+    default_status = Status.query.filter(Status.domain_id == project_domain.id,
+                                         Status.default_for_domain == True).first()
+    form.academic_year.choices = [(y.year, y.year) for y in AcademicYear.query.order_by(text('start_date'))]
+    if form.validate_on_submit():
+        skills_required = form.skills_required.data
+        project = Project(title = form.title.data,
+                          client_id = current_user.id,
+                          overview = form.overview.data,
+                          deliverables = form.deliverables.data,
+                          resources = form.resources.data,
+                          academic_year = form.academic_year.data,
+                          status_id = default_status.id)
+
+        try:
+            db.session.add(project)
+            db.session.commit()
+            db.session.refresh(project)
+            if project.id:
+                for skill in skills_required:
+                    skill_required = SkillRequired(project_id=project.id, skill_id=skill)
+                    db.session.add(skill_required)
+                db.session.commit()
+                logger.info('NEW STUDENT PROJECT: {:d}, {}, {:d}'.format(project.id, project.title, project.client_id))
+                flash('You have successfully added a new project.')
+                return redirect(url_for('student.projects'))
+            else:
+                raise RuntimeError
+
+        except RuntimeError:
+            flash('An unknown problem occurred while trying to save a new project', 'error')
+            logger.error('PROJECT INSERT ERROR: {:d}, {}'.format(project.client_id, project.title))
+        except:
+            flash('There was a problem saving your project', 'error')
+
+    return render_template('student/projects/edit_project.html', add_project=add_project,
+                           form=form, title='Add project')
+
+
+@student.route('/edit_project/<int:id>', methods=['GET', 'POST'])
+@login_required
+def edit_project(id):
+    add_project = False
+
+    project = Project.query.get_or_404(id)
+    check_client(project)
+
+    form = ProjectForm(obj=project)
+    form.academic_year.choices = [(y.year, y.year) for y in AcademicYear.query.order_by(text('start_date'))]
+    form.skills_required.choices = [(s.id, s.name) for s in Skill.query.order_by(text('name'))]
+    form.skills_required.data = [s.skill_id for s in project.skills_required]
+    if form.validate_on_submit():
+        project.title = form.title.data
+        project.overview = form.overview.data
+        project.deliverables = form.deliverables.data
+        project.resources = form.resources.data
+        project.academic_year = form.academic_year.data
+        project.updated_date = datetime.now()
+        db.session.add(project)
+        SkillRequired.query.filter(SkillRequired.project_id==project.id).delete()
+        for skill in form.skills_required.raw_data:
+            if skill not in [s.skill_id for s in project.skills_required]:
+                skill_required = SkillRequired(project_id=project.id, skill_id=skill)
+                db.session.add(skill_required)
+        db.session.commit()
+        flash('You have successfully edited the project.')
+
+        return redirect(url_for('student.projects'))
+
+    return render_template('student/projects/edit_project.html', add_project=add_project,
+                           project=project, form=form, title="Edit project")
+
+
+@student.route('/projects/project_transition/<int:id>/<int:status_id>', methods=['GET', 'POST'])
+@login_required
+def transition(id, status_id):
+
+    project = Project.query.get_or_404(id)
+    check_client(project)
+
+    status = Status.query.get(status_id)
+    project.status_id = status.id
+    db.session.add(project)
+    db.session.commit()
+    flash('The project status is now ' + status.name)
+
+    return redirect(url_for('student.project', id=id))
 
 
 @student.route('/interest/add/<int:id>', methods=['GET', 'POST'])
@@ -112,6 +222,8 @@ def add_interest(id):
 @login_required
 def delete_interest(project_id, id):
     interest = Interest.query.get(id)
+    check_owner(interest)
+
     db.session.delete(interest)
     db.session.commit()
     flash('Interest deleted')
@@ -143,6 +255,8 @@ def create_team(project_id):
             db.session.refresh(team)
             team_member = TeamMember(user_id=current_user.id, team_id=team.id)
             db.session.add(team_member)
+            delete_notes(current_user.id, project_id)
+            delete_flags(current_user.id, project_id)
             db.session.commit()
             flash('Team created')
         except:
@@ -164,39 +278,68 @@ def edit_team(id):
     settings = Settings.query.first()
 
     team = Team.query.get(id)
+    check_member(team)
     form = TeamForm(obj=team)
-    # form.comment.data = team.comment
-    # form.vacancies.data = team.vacancies
 
     if form.validate_on_submit():
+        error = False
         team.comment=form.comment.data
         team.vacancies=form.vacancies.data
-        db.session.add(team)
 
-        if form.matric.data is not None:
+        if form.matric.data != '':
+            members = TeamMember.query.filter(TeamMember.team_id == team.id).all()
+            if len(members) < settings.maximum_team_size:
+                try:
+                    member = User.query.filter(User.username==form.matric.data).first()
+                    if member.id in [m.user_id for m in members]:
+                        flash('Already on the team', 'error')
+                        error=True
+                    else:
+                        team_member = TeamMember(user_id=member.id, team_id=team.id)
+                        db.session.add(team_member)
+                        delete_notes(member.id, team.project_id)
+                        delete_flags(member.id, team.project_id)
+                except AttributeError:
+                    flash('Matric number not recognised', 'error')
+                    error=True
+            else:
+                flash("Your team is already at the maximum size", "error")
+                error = True
+
+        if not error:
             try:
-                member = User.query.filter(User.username==form.matric.data)
-                team_member = TeamMember(user_id=member.id, team_id=team.id)
-                db.session.add(team_member)
-            except NoResultFound:
-                flash('Matric number not recognised', 'error')
-            except AttributeError:
-                pass
+                db.session.commit()
+                flash('Team updated')
+            except:
+                flash('There was a problem updating the team', 'error')
 
-        try:
-            db.session.commit()
-            flash('Team updated')
-        except:
-            flash('There was a problem updating the team', 'error')
-
-        return redirect(url_for('student.edit_team', id=team.id))
+            return redirect(url_for('student.edit_team', id=team.id))
 
     return render_template('student/teams/team.html',
                            settings=settings,
                            add_team=add_team,
                            team=team,
                            form=form,
-                           title='Edit team')
+                           title='Manage bid')
+
+
+@student.route('/member/pm/<int:team_id>/<int:id>', methods=['GET', 'POST'])
+@login_required
+def make_pm(team_id, id):
+    members = TeamMember.query.filter(TeamMember.team_id==team_id).all()
+    team = Team.query.get(team_id)
+    check_member(team)
+
+    for member in members:
+        if member.id == id:
+            member.project_manager = True
+        else:
+            member.project_manager = False
+
+    db.session.commit()
+    flash('Project manager updated')
+
+    return redirect(url_for('.edit_team', id=team_id))
 
 
 @student.route('/member/delete/<int:id>', methods=['GET', 'POST'])
@@ -204,6 +347,7 @@ def edit_team(id):
 def delete_member(id):
     team_member = TeamMember.query.get(id)
     team = team_member.team
+    check_member(team)
     if len(team.members) == 1:
         return redirect(url_for('student.delete_team', id=team.id))
 
@@ -218,12 +362,24 @@ def delete_member(id):
 @login_required
 def delete_team(id):
     team = Team.query.get(id)
+    check_member(team)
     project_id = team.project_id
     db.session.delete(team)
     db.session.commit()
     flash('Team deleted')
 
     return redirect(url_for('.project', id=project_id))
+
+
+@student.route('/team/preview/<int:id>', methods=['GET', 'POST'])
+@login_required
+def preview(id):
+    team = Team.query.get(id)
+    check_member(team)
+
+    return render_template('student/teams/preview.html',
+                           team=team,
+                           title='Preview')
 
 
 @student.route('/flag/set/<int:id>', methods=['GET', 'POST'])
@@ -256,6 +412,7 @@ def set_flag(id):
 @login_required
 def edit_flag(id):
     flag = Flag.query.get(id)
+    check_owner(flag)
     form = FlagForm(obj=flag)
     add_flag = False
 
@@ -281,6 +438,7 @@ def edit_flag(id):
 @login_required
 def delete_flag(id):
     flag = Flag.query.get(id)
+    check_owner(flag)
     project_id = flag.project_id
     db.session.delete(flag)
     db.session.commit()
@@ -292,7 +450,58 @@ def delete_flag(id):
 @student.route('/vacancies', methods=['GET', 'POST'])
 @login_required
 def vacancies():
-    teams = Team.query.filter(Team.vacancies != None).all()
+    teams = Team.query.filter(Team.vacancies != '').all()
     return render_template('student/projects/vacancies.html', teams=teams, title="Vacancies")
 
+
+#--------  Bid management ----------#
+
+
+@student.route('/student/complete/<int:id>', methods=['GET', 'POST'])
+@login_required
+def complete_team(id):
+
+    settings = Settings.query.first()
+    team = Team.query.get_or_404(id)
+    check_client(team.project)
+    if check_members(team, settings):
+        accepted = get_status('team', 'Accepted')
+        team.status_id = accepted.id
+        for member in TeamMember.query.filter(TeamMember.team_id==team.id):
+            delete_notes(member.id)
+            delete_flags(member.id)
+        db.session.commit()
+        flash('The team has been marked complete')
+        return redirect(url_for('student.project', id=team.project.id))
+    else:
+        flash('Your team needs at least {:d} members'.format(settings.minimum_team_size), 'error')
+        return redirect(url_for('student.edit_team', id=id))
+
+
+@student.route('/student/submit/<int:id>', methods=['GET', 'POST'])
+@login_required
+def submit(id):
+
+    settings = Settings.query.first()
+    team = Team.query.get_or_404(id)
+    check_member(team)
+
+    if not check_members(team, settings):
+        flash('Your team needs at least {:d} members'.format(settings.minimum_team_size), 'error')
+        return redirect(url_for('student.edit_team', id=id))
+
+    if not check_programmes(team):
+        flash('Team members must be from more than one programme', 'error')
+        return redirect(url_for('student.edit_team', id=id))
+
+    if not check_pm(team):
+        flash('Your team must have a project manager', 'error')
+        return redirect(url_for('student.edit_team', id=id))
+
+    submitted = get_status('team', 'Submitted')
+    team.status_id = submitted.id
+    alert(team.project.client_id, 'submit', project=team.project)
+    db.session.commit()
+    flash('Your bid has been submitted')
+    return redirect(url_for('student.project', id=team.project.id))
